content/media/test/test_bug493187.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=493187
-->

<head>
  <title>Bug 493187 - enter HAVE_FUTURE_DATA when seeking within buffered data even if new data isn't arriving</title>
  <script type="text/javascript" src="/MochiKit/packed.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="use_large_cache.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=493187">Mozilla Bug 493187</a>
<pre id="test">
<script class="testbody" type="text/javascript">
var v;
var startedSeek = false;
var completed = false;

function start() {
  v = document.getElementById('v');
  v.currentTime = 1.0;
}

function startSeeking() {
  startedSeek = true;
}

function canPlayThrough() {
  if (startedSeek && !completed) {
    ok(true, "Got canplaythrough after seek");
    completed = true;
    SimpleTest.finish();
  }
}

SimpleTest.waitForExplicitFinish();
</script>
</pre>
<video id='v' src='seek.ogv' oncanplaythrough='canPlayThrough()' onseeking='startSeeking()'
       autobuffer onloadedmetadata='start()'></video>
</body>
</html>
